update tests: integrate with pytest; fix numba/pytest related issues
