+++ tests/test_neighbors.py
@@ -1,5 +1,6 @@
 from dynamo.tools.connectivity import _gen_neighbor_keys, check_and_recompute_neighbors, check_neighbors_completeness
-from utils import *
+
+# import utils
 import networkx as nx
 import dynamo as dyn
 import matplotlib.pyplot as plt
@@ -49,7 +50,7 @@ def test_neighbors_no_pca_key():
 
 if __name__ == "__main__":
     # generate data if needed
-    adata = gen_or_read_zebrafish_data()
+    adata = utils.gen_or_read_zebrafish_data()
     test_neighbors_subset(adata)
     test_broken_neighbors_check_recompute(adata)
     test_neighbors_no_pca_key()
