+++ tests/test_space.py
@@ -1,9 +1,8 @@
-from utils import *
+# import utils
 
+import pytest
 import dynamo as dyn
 from dynamo.pl import space
-
-logger = LoggerManager.get_main_logger()
 
 
 def test_space_plot_simple1(adata, color=["Cell_type"]):
@@ -14,8 +13,9 @@ def test_space_plot_simple1(adata, color=["Cell_type"]):
     space(adata, color=color, marker="*", save_show_or_return="show")
 
 
-def test_space_stack_color(adata):
-    adata = read_test_spatial_genomics_data()
+@pytest.mark.skip(reason="todo: add test data for spatial genomics")
+def test_space_stack_color(adata, utils):
+    adata = utils.read_test_spatial_genomics_data()
     genes = adata.var_names[:18]
     # space(adata, genes=genes, marker="*", save_show_or_return="show", stack_genes=False)
     space(
@@ -29,13 +29,14 @@ def test_space_stack_color(adata):
     )
 
 
+@pytest.mark.skip(reason="todo: add test data for spatial genomics")
 def test_space_data():
     adata = dyn.read_h5ad("allstage_splice.h5ad")
 
 
 if __name__ == "__main__":
     # generate data if needed
-    adata = gen_or_read_zebrafish_data()
+    adata = utils.gen_or_read_zebrafish_data()
 
     # TODO use a fixture in future
     test_space_plot_simple1(adata)
